requirements.txt
streamlit
openai
langchain
pandas
python-dotenv